Replaced the src directory with the version in https://github.com/nimetu/docky It appears to be more compatible and may fix the crashing on character creation

+++ src/ai_service/script_vm.cpp
@@ -991,7 +991,7 @@ void CScriptVM::interpretCode(
 					{
 						float val = (float&)stack.top();
 						stack.pop();
-						compValue = *((uint32*)&val);
+						compValue = *((size_t*)&val);
 					}
 					break;
 				default:
